Add tests & fix doctrine config

+++ tests/Admin/PortalElementAdminTest.php
@@ -12,10 +12,10 @@
 namespace Tests\Admin;
 
 use Agate\Entity\PortalElement;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class PortalElementAdminTest extends AbstractEasyAdminTest
 {
-    private $formEntityFieldName;
     private $file;
 
     /**
@@ -34,21 +34,86 @@ class PortalElementAdminTest extends AbstractEasyAdminTest
         return PortalElement::class;
     }
 
-    protected function setUp()
-    {
-        parent::setUp();
-        if (!$this->formEntityFieldName) {
-            $this->formEntityFieldName = strtolower($this->getEntityName());
-        }
-        $this->file = tempnam(sys_get_temp_dir(), 'portal_element_admin_test');
-    }
-
     protected function tearDown()
     {
         parent::tearDown();
-        if (file_exists($this->file)) {
+
+        if ($this->file !== null && file_exists($this->file)) {
             unlink($this->file);
         }
+    }
+
+    public function testNewValidFileUpload()
+    {
+        static::resetDatabase();
+
+        $this->createImage();
+
+        $submitted = [
+            'portal' => 'agate',
+            'locale' => 'en',
+            'image' => new UploadedFile($this->file, 'uploaded_file.jpg'),
+            'title' => 'Portail Esteren',
+            'subtitle' => 'sub',
+            'buttonText' => 'button',
+            'buttonLink' => '/',
+        ];
+
+        $expected = [
+            'portal' => 'agate',
+            'locale' => 'en',
+            'title' => 'Portail Esteren',
+            'subtitle' => 'sub',
+            'buttonText' => 'button',
+            'buttonLink' => '/',
+        ];
+
+        /** @var PortalElement $entity */
+        $entity = $this->submitData($submitted, $expected, 'new');
+
+        static::assertRegExp('~^portal_element_\w+-tmp_pe_[a-z0-9]+\.[0-9]+\.jpeg$~isUu', $entity->getImageUrl(), $entity->getImageUrl());
+
+        $filePath = static::$kernel->getContainer()->getParameter('kernel.project_dir').'/public/uploads/'.$entity->getImageUrl();
+
+        static::assertFileExists($filePath);
+
+        unlink($filePath);
+    }
+
+    public function testEditValidFileUpload()
+    {
+        static::resetDatabase();
+
+        $this->createImage();
+
+        $submitted = [
+            'image' => new UploadedFile($this->file, 'uploaded_file.jpg'),
+            'title' => 'Portail Esteren',
+            'subtitle' => 'sub',
+            'buttonText' => 'button',
+            'buttonLink' => '/',
+        ];
+
+        $expected = [
+            'id' => 1,
+            'portal' => 'esteren',
+            'locale' => 'fr',
+            'title' => 'Portail Esteren',
+            'subtitle' => 'sub',
+            'buttonText' => 'button',
+            'buttonLink' => '/',
+        ];
+
+        /** @var PortalElement $entity */
+        $entity = $this->submitData($submitted, $expected, 'edit');
+
+        static::assertRegExp('~^portal_element_\w+-tmp_pe_[a-z0-9]+\.[0-9]+\.jpeg$~isUu', $entity->getImageUrl(), $entity->getImageUrl());
+
+        $filePath = static::$kernel->getContainer()->getParameter('kernel.project_dir').'/public/uploads/'.$entity->getImageUrl();
+
+        static::assertFileExists($filePath);
+
+        unlink($filePath);
     }
 
     /**
@@ -81,5 +146,36 @@ class PortalElementAdminTest extends AbstractEasyAdminTest
     public function provideEditFormData()
     {
         return false;
+    }
+
+    private function createImage(int $width = 1900, $height = 900)
+    {
+        $this->file = tempnam(sys_get_temp_dir(), 'portal_test').'.jpg';
+
+        imagejpeg(imagecreate($width, $height), $this->file);
+    }
+
+    protected static function resetDatabase()
+    {
+        parent::resetDatabase();
+
+        $kernel = static::bootKernel();
+
+        $class = PortalElement::class;
+
+        $kernel->getContainer()->get('doctrine.orm.entity_manager')
+            ->createQuery(<<<DQL
+                DELETE FROM {$class} element 
+                WHERE element.portal = :portal 
+                AND element.locale = :locale
+DQL
+)
+            ->setParameter('portal', 'agate')
+            ->setParameter('locale', 'en')
+            ->execute()
+        ;
+
+        static::ensureKernelShutdown();
+        static::$kernel = null;
     }
 }
